Move OsmFilter and make it package-private
