+++ src/main/java/org/opentripplanner/openstreetmap/model/OSMRelation.java
@@ -56,6 +56,11 @@ public class OSMRelation extends OSMWithTags {
     return isPublicTransport() && isTag("public_transport", "stop_area");
   }
 
+  @Override
+  public boolean isRoutable() {
+    return super.isRoutable() && OsmChecks.isRoutable(this);
+  }
+
   private boolean isType(String type) {
     return isTag("type", type);
   }
