+++ src/main/java/org/opentripplanner/openstreetmap/model/OSMWay.java
@@ -83,6 +83,11 @@ public class OSMWay extends OSMWithTags {
     return "osm way " + id;
   }
 
+  @Override
+  public boolean isRoutable() {
+    return super.isRoutable() && OsmChecks.isRoutable(this);
+  }
+
   /**
    * Returns true if way geometry is a closed loop
    */
